samples/mcp-string-server/pyproject.toml
[project]
name = "mcp-string-server"
version = "0.0.1"
description = "String Operations MCP Server"
authors = [{ name = "John Doe" }]
dependencies = [
    "uipath-mcp>=0.1.1",
]
requires-python = ">=3.11"
